Rename messages app name and add migrations
