+++ backend/core/urls.py
@@ -26,7 +26,7 @@ urlpatterns = [
     path('rooms/', include('rooms.urls')),
     path('reservations/', include('reservations.urls')),
     path('payments/', include('payments.urls')),
-    path('messages/', include('messages.urls')),
+    path('messages/', include('messages_.urls')),
     path('schema/', SpectacularAPIView.as_view(), name='schema'),
     path('swagger/', SpectacularSwaggerView.as_view(url_name='schema'), name='swagger-ui'),
     path('admin/', admin.site.urls),
